refactor code by moving hooks to dedecated files and isolating error boundarys to designated components
