+++ src/app/profile/[id]/components/content/cart/cart-panel.tsx
@@ -1,0 +1,80 @@
+import { CartType } from "@/lib/types";
+import { ErrorBoundary } from "react-error-boundary";
+import {
+  Suspense,
+  useEffect,
+  useState,
+} from "react";
+import BrandButton from "@/components/ui/button";
+import useCarts, { clearCartCache } from "@/lib/hooks/useCarts";
+
+function ErrorFallback({
+  error,
+  resetErrorBoundary,
+}: {
+  error: Error;
+  resetErrorBoundary: () => void;
+}) {
+  const status =
+    error.cause && typeof error.cause === "number" ? error.cause : 400;
+
+  return (
+    <div role="alert" className="flex flex-col gap-4 p-6">
+      {status === 404 ? (
+        <div className="flex flex-col justify-start w-full gap-2 text-brand-primary">
+          <h1 className="text-3xl font-bold">
+            You don&apos;t have any cart drops right now!
+          </h1>
+          <p className="text-lg">
+            Manage your drops by heading to the shopping cart tab above.
+          </p>
+          <hr />
+        </div>
+      ) : (
+        <div className="flex items-start justify-between w-full text-brand-accent">
+          <div className="flex flex-col gap-1">
+            <h1 className="text-3xl font-bold">Something went wrong!</h1>
+            <pre className="text-md">{error.message}</pre>
+          </div>
+          <BrandButton variant="delete" onClick={resetErrorBoundary}>
+            Try Again
+          </BrandButton>
+        </div>
+      )}
+    </div>
+  );
+}
+
+function ProfileCartInner({
+  userId,
+}: {
+  userId: string;
+}) {
+  const { data } = useCarts(userId);
+  const [carts, setCarts] = useState<CartType[] | null>(null);
+
+  useEffect(() => {
+    setCarts(data);
+  }, [data]);
+
+  return (
+    <div className="flex flex-col gap-4 p-6">
+      {(carts ?? []).map((cart) => (
+        <div key={cart.id}>{JSON.stringify(cart)}</div>
+      ))}
+    </div>
+  );
+}
+export default function ProfileCart({ userId }: { userId: string }) {
+  return (
+    <ErrorBoundary
+      FallbackComponent={ErrorFallback}
+      onReset={() => clearCartCache(userId)}
+      resetKeys={[userId]}
+    >
+      <Suspense>
+        <ProfileCartInner userId={userId} />
+      </Suspense>
+    </ErrorBoundary>
+  );
+}
