+++ src/app/profile/[id]/components/content/delete/delete-panel.tsx
