+++ src/app/profile/[id]/components/content/details/details-skeleton.tsx
